+++ libs/server/best-scores/src/lib/best-subcompetency-score/entities/best-subcompetency-score.entity.ts
@@ -8,7 +8,7 @@ export class BestSubCompetencyScoreEntity {
   @PrimaryGeneratedColumn()
   id!: number;
 
-  @Column({ type: 'integer', nullable: false })
+  @Column({ type: 'numeric', nullable: false })
   best_score_absolute!: number;
 
   @Column({ type: 'numeric', nullable: false })
